Added the implementation of the CategoriesManager and improved the Sqlite backend.

+++ src/it/robol/android/piggybank/data/Account.java
@@ -6,7 +6,7 @@ public class Account {
 	
 	private DataProvider mProvider = null;
 	
-	public int id = -1;
+	public long id = -1;
 	public String name = "New Account";
 	public String state = "active";
